requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
python-dotenv>=1.0
matplotlib>=3.8
pvlib>=0.10
citylearn>=2.5.0
gymnasium<=0.28.1
torch>=2.0
stable-baselines3>=2.3.0
tabulate>=0.9.0
